@@ -3,28 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.8.0",
+    "@reduxjs/toolkit": "^1.9.3",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^24.9.1",
-    "@types/node": "^12.20.46",
-    "@types/react": "^16.14.23",
-    "@types/react-dom": "^16.9.14",
-    "@types/react-redux": "^7.1.23",
     "node-sass": "^6.0.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-alert": "^7.0.3",
     "react-alert-template-basic": "^1.0.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-final-form": "^6.5.9",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "4.0.3",
-    "typescript": "~4.1.5"
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "export PORT=3003 && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -45,6 +40,7 @@
     ]
   },
   "devDependencies": {
+    "@types/jest": "^29.5.5",
     "@types/react-alert": "^7.0.2",
     "react-error-overlay": "^6.0.9"
   }
