@@ -1,5 +1,6 @@
-export const serverApiBaseUrl = 'http://localhost:8080/api';
+export const serverApiBaseUrl = 'http://localhost:3000/api';
 export const serverWsApiUrl = 'ws://localhost:8080/api';
+export const websocketServer = 'ws://localhost:3001';
 export const binanceApiBaseUrl = 'https://api.binance.com';
 export const binanceWsApiUrl = 'wss://stream.binance.com:9443';
 export const binanceFutApiBaseUrl = 'https://fapi.binance.com';
